normalize score map, add test program to demonstrate output

+++ GeneticMain/BoardGenerator.h
@@ -8,7 +8,7 @@ public:
     BoardGenerator();
     void generateBoard(int[][3], int, bool playerOneFirst=true);
     void generateBoardDef(int[][3]);
-    void generateScoreMap(int boardIn[][3], int scoreBoard[][3], int player);
+    void generateScoreMap(int boardIn[][3], float scoreBoard[][3], int player);
 private:
     void genBoard(int[][3], int, bool);
 };
